+++ packages/office-ui-fabric-react/src/components/Pivot/examples/Pivot.Basic.Example.tsx
@@ -3,10 +3,21 @@ import { Label } from 'office-ui-fabric-react/lib/Label';
 import { BaseComponent } from '../../../Utilities';
 import { ActionButton } from '../../Button';
 import { IPivotProps } from '../Pivot.types';
+import { IPivotItemProps } from '../PivotItem.types';
 // import { IPivotItemProps } from '../PivotItem.types';
 import { Pivot, PivotItem, IPivotState } from 'office-ui-fabric-react/lib/Pivot';
 import * as exampleStylesImport from '../../../common/_exampleStyles.scss';
 const exampleStyles: any = exampleStylesImport;
+
+export class MyCustomPivotItem extends BaseComponent<IPivotItemProps, {}> {
+  public render() {
+    return (
+      <aside { ...this.props.with }>
+        <Label>Special Pivot</Label>
+      </aside>
+    );
+  }
+}
 
 export class PivotBasicExample extends React.Component<any, any> {
   public render() {
@@ -22,10 +33,18 @@ export class PivotBasicExample extends React.Component<any, any> {
           <PivotItem linkText='My Files'>
             <Label className={ exampleStyles.exampleLabel }>Pivot #1</Label>
           </PivotItem>
-          <PivotItem linkText='Recent'>
+          <PivotItem
+            as={ MyCustomPivotItem }
+            with={ { 'data-name': 'third' } }
+            linkText='Recent'
+          >
             <Label>Pivot #2</Label>
           </PivotItem>
-          <PivotItem linkText='Shared with me'>
+          <PivotItem
+            as={ 'span' }
+            with={ { 'data-name': 'third' } }
+            linkText='Shared with me'
+          >
             <Label>Pivot #3</Label>
           </PivotItem>
         </Pivot>
